clear warning and sprites

+++ ThePlatformer/BaseGameObject.cpp
@@ -362,7 +362,7 @@ namespace GameObjects {
 		if (invicible)
 			return;
 
-		this->life -= std::round(damageMsg->damage / this->powerupDefense);
+		this->life -= (int) std::round(damageMsg->damage / this->powerupDefense);
 
 		if (damageMsg->damageType == GameMessage::DamageMessage::DASH) {
 			GameComponents::TimerComponent* timer = new GameComponents::TimerComponent(this);
